pNFS Refactor
Move pNFS functionality back for common use.  Clean up configure,
automake, and conditional compile options.  Within the C code there are three defines:

 _PNFS: Some pNFS functionality has been enabled, for common code.
 _PNFS_MDS: Metadata server functionality has been enabled
 _PNFS_DS: Data server functionality has been enabled

And three corresponding automake variables

 USE_PNFS, USE_PNFS_MDS, and USE_PNFS_DS

In order to let the Shell compile cleanly without pulling in the
entire NFS daemon, pNFS functionality has been split out from other
FSAL functionality.  Be sure to set FSALPNFS_LIB for your FSAL to get
it to link in.

Not yet done:

  The PARALLELL_FS code has not been refactored to work with the rest
  of the pNFS system.

  The configuration options to disable or enable pNFS support on a
  per-export basis need to be looked at.

  This probably does not work with a shared FSAL.  I was going to wait
  until after Jim Lieb's refactor branch hit mainline to address that.

+++ src/FSAL/FSAL_CEPH/fsal_mds.c
@@ -76,7 +76,7 @@ CEPHFSAL_layoutget(fsal_handle_t *exthandle,
         the Ceph cluster. */
      struct ceph_file_layout file_layout;
      /* Width of each stripe on the file */
-     uint64_t stripe_width = 0;
+     uint32_t stripe_width = 0;
      /* Utility parameter */
      nfl_util4 util = 0;
      /* The last byte that can be accessed through pNFS */
